requirements.txt
fastapi==0.78.0
uvicorn==0.17.6
python-dotenv==0.20.0
sentence-transformers==2.2.0
